requirements.txt
click==7.1.2
Faker==4.14.2
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
psycopg2-binary==2.8.6
python-dateutil==2.8.1
redis==3.5.3
rq==1.6.1
six==1.15.0
text-unidecode==1.3
Werkzeug==1.0.1
